✨ Player stuck detection
Detect if a player is stuck in a tilemap and respawn them if this is case. In future it might be a good idea to have some sort of notification for why you get respawned in this situation, or even have a HP / suffocation system so  players are not frustrated when accidentally getting themselves stuck.